added homepage moved previous customer list page to own route
