+++ app/globals.css
@@ -122,3 +122,29 @@
     @apply bg-background text-foreground;
   }
 }
+
+@layer utilities {
+  .bg-svg-top::before,
+  .bg-svg-bottom::after {
+    content: "";
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-repeat: no-repeat;
+    background-size: contain;
+    z-index: 0;
+    pointer-events: none;
+  }
+
+  .bg-svg-top::before {
+    top: 0;
+    background-position: top;
+    background-image: url("data:image/svg+xml,%3Csvg width='397' height='289' viewBox='0 0 397 289' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M173 133.5C92 195.5 5.5 216.5 -3.5 277C-12.5 337.5 -81.5 133.5 -81.5 133.5L1.22449 -34.0395C-51.1376 -62.2455 57 -147 57 -147L1.22449 -34.0395C24.1966 -21.665 78.0605 -20.1746 185.5 -41.5C537.835 -111.434 378.278 23.0159 220 57C61.7221 90.9841 254 71.5 173 133.5Z' fill='%23CF2395'/%3E%3Cpath d='M143.07 121.824C73.6647 174.391 -0.453453 192.196 -8.16516 243.492C-15.8769 294.787 -75 121.824 -75 121.824L-4.11696 -20.2256C-48.9838 -44.1402 43.6747 -116 43.6747 -116L-4.11696 -20.2256C15.5669 -9.73381 61.7205 -8.47015 153.781 -26.551C455.681 -85.845 318.964 28.1493 183.342 56.963C47.7209 85.7766 212.475 69.2569 143.07 121.824Z' fill='%23993377'/%3E%3Cpath d='M109.922 112.356C46.6112 161.504 -20.999 178.151 -28.0336 226.11C-35.0682 274.069 -89 112.356 -89 112.356L-24.3408 -20.4548C-65.2681 -42.814 19.2545 -110 19.2545 -110L-24.3408 -20.4548C-6.38535 -10.6454 35.7158 -9.46397 119.693 -26.3688C395.085 -81.8063 270.372 24.7737 146.659 51.7133C22.9454 78.653 173.234 63.2077 109.922 112.356Z' fill='%23772F5F'/%3E%3C/svg%3E");
+  }
+
+  .bg-svg-bottom::after {
+    bottom: 0;
+    background-position: bottom;
+    background-image: url("data:image/svg+xml,%3Csvg width='395' height='300' viewBox='0 0 395 300' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M235.296 174.53C311.254 106.446 395.869 78.814 400.159 17.7989C404.449 -43.2163 489.033 154.831 489.033 154.831L419.525 328.271C473.913 352.339 372.66 445.21 372.66 445.21L419.525 328.271C395.664 317.712 341.846 320.395 236.38 349.973C-109.485 446.969 39.1862 300.572 194.359 254.439C349.531 208.305 159.338 242.614 235.296 174.53Z' fill='%23CF2395'/%3E%3Cpath d='M266.04 183.854C331.169 126.073 403.686 102.584 407.404 50.8457C411.122 -0.892758 483.456 166.975 483.456 166.975L423.781 314.085C470.364 334.455 383.546 413.271 383.546 413.271L423.781 314.085C403.344 305.148 357.231 307.461 266.847 332.613C-29.5584 415.098 97.9251 290.863 230.909 251.638C363.894 212.413 200.912 241.636 266.04 183.854Z' fill='%23993377'/%3E%3Cpath d='M299.822 190.728C359.139 136.827 425.257 114.997 428.559 66.6372C431.86 18.2774 498.147 175.331 498.147 175.331L443.962 312.748C486.497 331.872 407.429 405.399 407.429 405.399L443.962 312.748C425.301 304.358 383.235 306.439 300.818 329.793C30.5437 406.381 146.633 290.467 267.89 254.033C389.146 217.598 240.504 244.63 299.822 190.728Z' fill='%23772F5F'/%3E%3C/svg%3E");
+  }
+}
