+++ app/layout.tsx
@@ -7,19 +7,18 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({
+const RootLayout = ({
   children,
 }: Readonly<{
   children: React.ReactNode;
-}>) {
+}>) => {
   return (
     <html lang="pt-br">
       <body>
-        <Header />
-        <main className="p-4">
-          {children}
-        </main>
+        {children}
       </body>
     </html>
   );
 }
+
+export default RootLayout;
